requirements.txt
selenium
openpyxl
requests
tkinter
pandas
boto3
python-dotenv
xlsxwriter